Move to audio source folder and update logging
